added functianality with userInfo
