+++ src/app.module.ts
@@ -1,14 +1,14 @@
  
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { AuthModule } from './auth/auth.module';
+import { AuthModule } from './auth/modules/auth.module';
 import { ConfigModule, ConfigService } from '@nestjs/config'
 import { async } from 'rxjs';
 import { configValidationSchema } from './config.schema';
-import { UserInfoModule } from './user-info/user-info.module';
-import { Industries } from './user-info/industires/industries.entity';
-import { Specialities } from './user-info/specialities/specialities.entity';
-import { UserInfo } from './user-info/user-info.entity';
+import { UserInfoModule } from './user-info/modules/user-info.module';
+import { Industries } from './user-info/repositories/industires/industries.entity';
+import { Specialities } from './user-info/repositories/specialities/specialities.entity';
+import { UserInfo } from './user-info/repositories/user-info/user-info.entity';
 
 @Module({
   imports: [
