+++ src/auth/controllers/auth.controller.ts
@@ -1,6 +1,6 @@
 import { Body, Controller, Post } from '@nestjs/common';
-import { AuthService } from './auth.service';
-import { AuthCredentialsDto } from './dto/auth-credentials.dto';
+import { AuthService } from '../services/auth.service';
+import { AuthCredentialsDto } from '../dto/auth-credentials.dto';
 
 @Controller('auth')
 export class AuthController {
