+++ src/auth/dto/get-user.decorators.ts
@@ -1,5 +1,5 @@
 import { createParamDecorator, ExecutionContext } from "@nestjs/common";
-import { User } from "../user.entity";
+import { User } from "../repositories/user.entity";
 
 export const GetUser = createParamDecorator((data, context: ExecutionContext):User => {
     const req = context.switchToHttp().getRequest();
