+++ src/auth/jwt/jwt.strategy.ts
@@ -5,8 +5,8 @@ import { PassportStrategy } from '@nestjs/passport'
 import { InjectRepository } from "@nestjs/typeorm";
 import { ExtractJwt, Strategy } from 'passport-jwt'
 import { JwtPayload } from "./jwt-payload.interface";
-import { User } from "./user.entity";
-import { UsersRepository } from "./users.repository";
+import { User } from "../repositories/user.entity";
+import { UsersRepository } from "../repositories/users.repository";
 
 @Injectable()
 export class JwtStrategy extends PassportStrategy(Strategy){
