+++ src/auth/modules/auth.module.ts
@@ -1,13 +1,13 @@
 
 import { Module } from '@nestjs/common';
-import { AuthService } from './auth.service';
-import { AuthController } from './auth.controller';
+import { AuthService } from '../services/auth.service';
+import { AuthController } from '../controllers/auth.controller';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { PassportModule } from '@nestjs/passport';
 import { JwtModule } from '@nestjs/jwt';
-import { JwtStrategy } from './jwt.strategy';
+import { JwtStrategy } from '../jwt/jwt.strategy';
 import { ConfigModule, ConfigService } from '@nestjs/config';
-import { UsersRepository } from './users.repository';
+import { UsersRepository } from '../repositories/users.repository';
 
 @Module({
   imports:[
