+++ src/auth/repositories/users.repository.ts
@@ -2,7 +2,7 @@ import { ConflictException, InternalServerErrorException } from "@nestjs/common"
 import { EntityRepository, Repository } from "typeorm";
 import { User } from "./user.entity";
 import * as bcrypt from 'bcrypt'
-import { AuthCredentialsDto } from "./dto/auth-credentials.dto";
+import { AuthCredentialsDto } from "../dto/auth-credentials.dto";
 
 @EntityRepository(User)
 export class UsersRepository extends Repository<User>{
