+++ src/payment/payment.controller.ts
@@ -8,5 +8,6 @@ export class PaymentController {
     @Post()
     payment(@Body('paymentCredentials') paymentCredentials:PaymentCredentials){
         const stripe = new Stripe('', {apiVersion: '2020-08-27'})
+        stripe.charges.create
     }
 }
