+++ src/user-info/controllers/user-info.controller.ts
@@ -1,10 +1,10 @@
 import { Body, Controller, Get, Param, Post, UseGuards } from '@nestjs/common';
 import { AuthGuard } from '@nestjs/passport';
 import { GetUser } from 'src/auth/dto/get-user.decorators';
-import { User } from 'src/auth/user.entity';
-import { UserCredentialsDto } from './dto/user-credentials.dto';
-import { UserInfo } from './user-info.entity';
-import { UserInfoService } from './user-info.service';
+import { User } from 'src/auth/repositories/user.entity';
+import { UserCredentialsDto } from '../dto/user-credentials.dto';
+import { UserInfo } from '../repositories/user-info/user-info.entity';
+import { UserInfoService } from '../services/user-info.service';
 
 @Controller('user-info')
 @UseGuards(AuthGuard())
@@ -17,7 +17,7 @@ export class UserInfoController {
     }
 
     @Get()
-    getUserInfo(@Param('id') id, @GetUser() user:User):Promise<UserInfo>{
-        return this.userInfoService.getUserInfo(id);
+    getUserInfo(@GetUser() user:User):Promise<UserInfo>{
+        return this.userInfoService.getUserInfo(user.id);
     }
 }
