+++ src/user-info/repositories/industires/industries.repository.ts
