+++ src/user-info/repositories/specialities/specialities.entity.ts
@@ -1,16 +1,12 @@
 import { Exclude } from "class-transformer";
 import { Column, Entity, ManyToOne, PrimaryGeneratedColumn } from "typeorm";
-import { UserInfo } from "../user-info.entity";
+import { UserInfo } from "../user-info/user-info.entity";
 
 @Entity()
-export class Industries{
+export class Specialities{
     @PrimaryGeneratedColumn('uuid')
     id:string;
 
     @Column()
     name:string;
-
-    @ManyToOne(_type => UserInfo, userInfo => userInfo.industries, {eager: false})
-    @Exclude({ toPlainOnly:true })
-    userInfo:UserInfo;
 }
